% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{write.bib}
\alias{write.bib}
\title{Generate a Bibtex File from Package Citations}
\usage{
write.bib(entry, file = "Rpackages.bib", append = FALSE, verbose = TRUE)
}
\arguments{
\item{entry}{a \code{\link{bibentry}} object or a character vector of package
names.}

\item{file}{output Bibtex file.}

\item{append}{logical. If \code{TRUE} content is appended to the file.}

\item{verbose}{a logical to toggle verbosity.}
}
\value{
the list of Bibtex objects -- invisibly.
}
\description{
Generates a Bibtex file from a \code{\link{bibentry}} object or a list of
packages. It is useful for adding relevant citations in Sweave documents.
}
\examples{
tmp <- tempfile(fileext = ".bib")
write.bib(c("bibtex", "utils", "tools"), file = tmp)
bibs <- read.bib(tmp)

tmp2 <- tempfile(fileext = ".bib")
write.bib(bibs, tmp2)
unname(tools::md5sum(tmp) == tools::md5sum(tmp2))

unlink(c(tmp, tmp2))
}
\references{
\emph{[R] Creating bibtex file of all installed packages?}
Achim Zeileis. R-help mailing list.
}
\author{
Renaud Gaujoux, based on the function \code{Rpackages.bib}
from Achim Zeileis (see \emph{References}).
}
