\name{bibtex-package}
\Rdversion{1.1}
\alias{bibtex-package}
\alias{bibtex}
\docType{package}
\title{
bibtex parser
}
\description{
parser for bibliography files written in bibtex format
}
\details{
\tabular{ll}{
Package: \tab bibtex\cr
Type: \tab Package\cr
Version: \tab 0.3-0\cr
Date: \tab 2011-11-16\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois

Maintainer: Romain Francois <romain@r-enthusiasts.com>
}
\references{
Nelson H. F. Beebe. bibparse 1.04. 1999. \url{http://www.math.utah.edu/~beebe}
}
\keyword{ package }
\seealso{
	\code{\link{read.bib}}
}

