% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex-package.R
\docType{package}
\name{bibtex-package}
\alias{bibtex}
\alias{bibtex-package}
\title{bibtex: Bibtex Parser}
\description{
Utility to parse a bibtex file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/bibtex}
  \item \url{https://docs.ropensci.org/bibtex/}
  \item Report bugs at \url{https://github.com/ropensci/bibtex/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [contributor]

Authors:
\itemize{
  \item Romain Francois \email{romain@purrple.cat} (\href{https://orcid.org/0000-0002-2444-4226}{ORCID}) (principal developer of bibtex R package) [copyright holder]
}

Other contributors:
\itemize{
  \item Scott Chamberlain \email{myrmecocystus@gmail.com} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID}) [contributor]
  \item Kurt Hornik \email{Kurt.Hornik@R-project.org} (\href{https://orcid.org/0000-0003-4198-9911}{ORCID}) [contributor]
  \item Brian Ripley \email{ripley@stats.ox.ac.uk} (compilation fixes) [contributor]
  \item Michael Koohafkan \email{michael.koohafkan@gmail.com} [contributor]
  \item Mathew W. McLean \email{mathew.w.mclean@gmail.com} (\href{https://orcid.org/0000-0002-7891-9645}{ORCID}) [contributor]
}

}
\keyword{internal}
