% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{do_read_bib}
\alias{do_read_bib}
\title{convenience wrapper around .External call}
\usage{
do_read_bib(file, encoding = "unknown", srcfile)
}
\arguments{
\item{file}{file name}

\item{encoding}{encoding}

\item{srcfile}{output of \code{\link{srcfile}}}
}
\description{
This is a convenience function for packages that do need to call the internal
functionality of \code{\link{read.bib}} but does different processing. This is
a simple wrapper around the \code{.External} code used by \code{\link{read.bib}}
}
\details{
The parser is greatly inspired from the \samp{bibparse} library.
}
\seealso{
\code{\link[utils]{bibentry}}
}
